Updated requirements/minor fixes
