+++ requirements.txt
@@ -1,1 +1,5 @@
-git+git://github.com/GPflow/GPflow.git
+tensorflow==1.13.1
+tensorflow-probability==0.6.0
+gpflow=1.5.0
+gym=0.15.3
+oct2py
